requirements.txt
beautifulsoup4==4.12.3
faunadb==4.5.1
Flask==2.1.2
Flask_Cors==3.0.10
Flask_Limiter==2.4.6
Jinja2==3.1.2
jsonschema==4.22.0
lxml==4.9.1
python_dateutil==2.9.0.post0
pytz==2021.3
requests==2.30.0
rss_parser==2.0.0
toml==0.10.2
Werkzeug==2.1.2
gunicorn==22.0.0